Decrease number of compilation errors on Scala 3
